Fix bugs initials

+++ src/domain/usecases/RemoveFile.mjs
@@ -7,8 +7,13 @@ export default class RemoveFile {
     async execute() {
         try {
             const filename = this.fileController.getFilename()
-            await this.fileRepository.removeFile(filename)
-            this.fileController.sendFileData({filename, removed: true})
+            const found = await this.fileRepository.exist(filename)
+            if (found) {
+                await this.fileRepository.removeFile(filename)
+                this.fileController.sendFileData({filename, removed: true})
+            } else {
+                this.fileController.sendNotFound()
+            }
         } catch (err) {
             this.fileController.sendError(err)
         }
